requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==21.7

# Google APIs
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
pytz==2024.2
python-dateutil==2.9.0

# For email parsing
beautifulsoup4==4.12.3
html2text==2024.2.26
